#6 폰트 수정및 파일 구조 변경
