+++ src/pages/_app.tsx
@@ -1,6 +1,7 @@
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
 import Navbar from "@/components/navbar";
+
 export default function App({ Component, pageProps }: AppProps) {
   return (
     <>
