+++ src/styles/globals.css
@@ -1,5 +1,7 @@
 body {
     margin: 0;
     padding: 0;
+    font-family: 'Noto Sans KR', sans-serif;
+ 
   }
